requirement.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.12.7
click==8.1.3
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.5
python-telegram-bot==13.15
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==2.0.5.post1
tornado==6.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
Werkzeug==2.2.3
zipp==3.15.0
